moved code folders, trying to add type checking and the right amound of abstraction
