+++ docs/motor_characterisation/magtrol_graphs/ServoFromDrury.csv
@@ -1,0 +1,180 @@
+x,Continuous,Intermittent,Limited
+0.058,2.5204,8.192,
+0.1508,2.5143,8.192,
+0.151,2.5082,8.192,
+0.2439,2.5021,8.192,
+0.2442,2.496,8.192,
+0.3369,2.4899,8.192,
+0.3373,2.4838,8.192,
+0.43,2.4777,8.192,
+0.4304,2.4716,8.192,
+0.5049,2.4655,8.192,
+0.523,2.4594,8.192,
+0.5979,2.4533,8.192,
+0.6161,2.4472,8.192,
+0.6911,2.4411,8.192,
+0.7092,2.435,8.192,
+0.7841,2.4289,8.192,
+0.8022,2.4228,8.192,
+0.8766,2.4167,8.192,
+0.8772,2.4106,8.192,
+0.9697,2.4045,8.192,
+0.9703,2.3984,8.192,
+1.0628,2.3923,8.192,
+1.0633,2.3862,8.192,
+1.1559,2.3801,8.192,
+1.1565,2.374,8.192,
+1.2303,2.3679,8.192,
+1.231,2.3618,8.192,
+1.3234,2.3557,8.192,
+1.324,2.3496,8.192,
+1.4164,2.3435,8.192,
+1.4172,2.3374,8.192,
+1.5095,2.3313,8.192,
+1.5102,2.3252,8.192,
+1.6025,2.3191,8.192,
+1.6033,2.313,8.192,
+1.6956,2.3069,8.192,
+1.6964,2.3008,8.192,
+1.7886,2.2947,8.192,
+1.7895,2.2886,8.192,
+1.8817,2.2825,8.192,
+1.8826,2.2764,8.192,
+1.9747,2.2703,8.192,
+1.9757,2.2642,8.192,
+2.0502,2.2581,8.192,
+2.0678,2.252,8.192,
+2.1433,2.2459,8.192,
+2.1608,2.2398,8.192,
+2.2364,2.2337,8.192,
+2.2539,2.2276,8.192,
+2.3294,2.2215,8.192,
+2.3469,2.2154,8.192,
+2.44,2.2093,8.192,
+2.4831,2.2032,8.192,
+2.5331,2.1971,8.192,
+2.6261,2.191,8.192,
+2.6367,2.1849,8.192,
+2.7192,2.1788,8.192,
+2.781,2.1727,8.192,
+2.8122,2.1666,8.192,
+2.9053,2.1605,8.192,
+2.9625,2.1544,8.192,
+2.9983,2.1483,8.192,
+3.0914,2.1422,8.192,
+3.1301,2.1361,8.192,
+3.1844,2.13,8.192,
+3.2231,2.1239,8.192,
+3.2775,2.1178,8.192,
+3.3162,2.1117,8.192,
+3.3705,2.1056,8.192,
+3.4093,2.0995,8.192,
+3.4636,2.0934,8.192,
+3.5024,2.0873,8.192,
+3.5566,2.0812,8.192,
+3.5955,2.0751,8.192,
+3.6497,2.069,8.192,
+3.6699,2.0629,8.192,
+3.7427,2.0568,8.192,
+3.7631,2.0507,8.192,
+3.8358,2.0446,8.192,
+3.8562,2.0385,8.192,
+3.9288,2.0324,8.192,
+3.9307,2.0263,8.192,
+4.0219,2.0202,8.1927,2.0202
+4.1895,2.003,8.1465,2.003
+4.2641,1.9934,8.1234,1.9934
+4.2705,1.9925,8.1236,1.9925
+4.3571,1.9764,8.1243,1.9764
+4.4503,1.952,8.0774,1.952
+4.508,1.9351,8.0596,1.9351
+4.5248,1.9303,8.0542,1.9303
+4.6181,1.906,8.0074,1.906
+4.6989,1.8832,7.9673,1.8832
+4.7113,1.8791,7.9605,1.8791
+4.7859,1.851,7.9135,1.851
+4.848,1.8249,7.8828,1.8249
+4.8791,1.8125,7.8666,1.8125
+4.9412,1.79,7.8265,1.79
+4.9537,1.7855,7.8196,1.7855
+5.0297,1.755,7.7837,1.755
+5.0469,1.7469,7.7727,1.7469
+5.1216,1.7088,7.7018,1.7088
+5.1229,1.7082,7.7007,1.7082
+5.1962,1.6623,7.6548,1.6623
+5.1975,1.6611,7.6539,1.6611
+5.2722,1.5902,7.5974,1.5902
+5.2895,1.5811,7.584,1.5811
+5.3641,1.5442,7.537,1.5442
+5.3654,1.5433,7.5365,1.5433
+5.4387,1.4858,7.4899,1.4858
+5.5333,1.4017,7.3957,1.4017
+5.6067,1.332,7.3244,1.332
+5.608,1.3308,7.3232,1.3308
+5.6813,1.2612,7.2774,1.2612
+5.6827,1.2598,7.2764,1.2598
+5.756,1.1902,7.2064,1.1902
+5.7574,1.1889,7.2052,1.1889
+5.8321,1.118,7.1476,1.118
+5.8493,1.0981,7.1357,1.0981
+5.8881,1.0469,7.1088,1.0469
+5.9629,0.9521,7.0548,0.9521
+5.9799,0.9308,7.0414,0.9308
+6.019,0.881,7.0067,0.881
+6.0546,0.8352,6.9705,0.8352
+6.0937,0.7862,6.9334,0.7862
+6.1292,0.7439,6.8996,0.7439
+6.1498,0.7151,6.8788,0.7151
+6.2039,0.6231,6.8287,0.6231
+6.2059,0.6201,6.8272,0.6201
+6.2785,0.5283,6.7816,0.5283
+6.2807,0.5253,6.7799,0.5253
+6.3368,0.4304,6.7256,0.4304
+6.3532,0.4003,6.7107,0.4003
+6.393,0.3354,6.6803,0.3354
+6.4465,0.2679,6.64,0.2679
+6.4491,0.2643,6.6378,0.2643
+6.5052,0.1693,6.5829,0.1693
+6.5212,0.1403,6.5691,0.1403
+6.5958,0.08,6.522,0.1
+6.6704,0,6.4511,
+6.7451,,6.3802,
+6.8198,,6.3093,
+6.913,,6.2624,
+6.9877,,6.1915,
+7.0624,,6.1206,
+7.137,,6.0735,
+7.2117,,6.0026,
+7.2863,,5.9317,
+7.3795,,5.8848,
+7.4542,,5.8139,
+7.5289,,5.743,
+7.6035,,5.696,
+7.6782,,5.6251,
+7.7529,,5.5541,
+7.8275,,5.5071,
+7.9021,,5.4362,
+7.9768,,5.3653,
+8.0701,,5.2945,
+8.1448,,5.2236,
+8.2194,,5.1766,
+8.2941,,5.1056,
+8.3874,,5.0349,
+8.462,,4.964,
+8.5367,,4.8931,
+8.6113,,4.846,
+8.742,,4.7279,
+8.8351,,4.7049,
+8.9285,,4.6103,
+9.0031,,4.5632,
+9.0963,,4.5164,
+9.1896,,4.4456,
+9.2643,,4.3747,
+9.3389,,4.3277,
+9.4322,,4.257,
+9.5068,,4.2099,
+9.6001,,4.1392,
+9.6747,,4.0921,
+9.7493,,4.0451,
+9.8425,,3.9982,
+9.9172,,3.9273,
